+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
 ## [Unreleased]
+- sed -i 's/MAX_NUM_OF_BITS/MAX_DATA_LENGTH/g' *
+- sed -i 's/N_xdigits/N_bytes/g' *
+- Reduce memory footprint by half
+- Deal with corner case, INT_MIN
+- Rewrite ebi(int)
+- sed -i 's/get_byte/get_byte/g' *
+- Add MACRO: #define EasyBigInteger ebi
+- Update README.md
+- sed -i 's/block/byte/g' *
+- Complete shift operators, operator<<(int) and operator>>(int)
+- Implement octadeciaml output, but not tested nor executed
+- Add macro BYTE_SIZE
 
 ## 2018.05.24
 - Check uint8_t exists by #ifndef UINT8_MAX